包装类型

+++ src/main/java/com/lianshang/generator/meta/ColumnTypeUtil.java
@@ -17,7 +17,7 @@ public class ColumnTypeUtil {
         }
         else if ("TINYINT".equalsIgnoreCase(type)
                 || "TINYINT UNSIGNED".equalsIgnoreCase(type)) {
-            return ColumnType.BYTE;
+            return ColumnType.INT;
         }
         else if ("INT".equalsIgnoreCase(type)
                 || "INT UNSIGNED".equalsIgnoreCase(type)) {
